[Venv] Install Venv
: Install Venv(Virtual Environment) which helps managing development
environment

python -m venv venv
: python: This is the Python interpreter.
-m: This option tells Python to run a module as a script.
venv: This is the name of the module to run. The venv module is a built-in module in Python 3 that provides support for creating and managing virtual environments.
venv: This is the name of the virtual environment directory that will be created. You can replace venv with the name of your choice.

source /path/to/venv/Scripts/activate
: Command for activating venv